pyproject.toml
[tool.ruff]
line-length = 100
include = ["*.py", "*.pyi"]
exclude = [
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    ".venv",
    "_build",
    "buck-out",
    "build",
    "dist",
]

select = ["E", "F", "I", "W", "N", "B", "C4", "UP", "RUF"]
ignore = ["E402"]


[tool.ruff.isort]
known-first-party = ["src", "gss_mol"]

[project]
name = "gss-mol"
version = "0.1.0"
requires-python = "== 3.12"
description = "GSS-Mol: Energy-guided diffusion sampling for molecular conformational search"

dependencies = [
"ase==3.25.0",
"rdkit",
"fairchem-core",
"numpy",
"scipy",
"omegaconf",

"torch==2.6.0+cu118; sys_platform == 'linux'",
"torch==2.6.0; sys_platform == 'darwin'",
"torchvision==0.21.0+cu118; sys_platform == 'linux'",
"torchvision==0.21.0; sys_platform == 'darwin'",
"torch_cluster",
"torch_geometric==2.5.3",
"torch_scatter",
"torch_sparse",
"torch_spline_conv",

"tqdm",
"matplotlib",
"e3nn",
"lightning",
"diffusers",
"wandb==0.2.2",
"mattersim==1.2.0",
"torchani",
"h5py==3.15.1",
"setuptools<82"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["gss_mol*", "src*"]

[project.scripts]
gss-mol-sample = "gss_mol.cli:sample"
gss-mol-relax = "gss_mol.cli:relax"
gss-mol-rss = "gss_mol.cli:rss"
gss-mol-plot = "gss_mol.cli:plot"

[tool.uv.sources]
torch = { index = "pytorch_linux",  marker = "sys_platform == 'linux'" }
torchvision = { index = "pytorch_linux",  marker = "sys_platform == 'linux'" }
pyg-lib = [
  { url = "https://data.pyg.org/whl/torch-2.6.0%2Bcu118/pyg_lib-0.4.0%2Bpt26cu118-cp312-cp312-linux_x86_64.whl",  marker = "sys_platform == 'linux'" },
  { url = "https://data.pyg.org/whl/torch-2.6.0%2Bcpu/pyg_lib-0.4.0%2Bpt26-cp312-cp312-macosx_14_0_universal2.whl",  marker = "sys_platform == 'darwin'" }
]
torch_cluster = [
  { url = "https://data.pyg.org/whl/torch-2.6.0%2Bcu118/torch_cluster-1.6.3%2Bpt26cu118-cp312-cp312-linux_x86_64.whl",  marker = "sys_platform == 'linux'" },
  { url = "https://data.pyg.org/whl/torch-2.6.0%2Bcpu/torch_cluster-1.6.3-cp312-cp312-macosx_10_13_universal2.whl",  marker = "sys_platform == 'darwin'"  }
]

torch_scatter = [
  { url = "https://data.pyg.org/whl/torch-2.6.0%2Bcu118/torch_scatter-2.1.2%2Bpt26cu118-cp312-cp312-linux_x86_64.whl",  marker = "sys_platform == 'linux'" },
  { url = "https://data.pyg.org/whl/torch-2.6.0%2Bcpu/torch_scatter-2.1.2-cp312-cp312-macosx_10_13_universal2.whl",  marker = "sys_platform == 'darwin'"  }
]

torch_sparse = [
  { url = "https://data.pyg.org/whl/torch-2.6.0%2Bcu118/torch_sparse-0.6.18%2Bpt26cu118-cp312-cp312-linux_x86_64.whl",  marker = "sys_platform == 'linux'" },
  { url = "https://data.pyg.org/whl/torch-2.6.0%2Bcpu/torch_sparse-0.6.18-cp312-cp312-macosx_11_0_universal2.whl",  marker = "sys_platform == 'darwin'"  }
]

torch_spline_conv = [
  { url = "https://data.pyg.org/whl/torch-2.6.0%2Bcu118/torch_spline_conv-1.2.2%2Bpt26cu118-cp312-cp312-linux_x86_64.whl",  marker = "sys_platform == 'linux'" },
  { url = "https://data.pyg.org/whl/torch-2.6.0%2Bcpu/torch_spline_conv-1.2.2-cp312-cp312-macosx_10_13_universal2.whl",  marker = "sys_platform == 'darwin'"  }
]

[tool.uv]
no-build-package = ["pyarrow"]

[[tool.uv.index]]
name = "pytorch_linux"
url = "https://download.pytorch.org/whl/cu118"
explicit = true
